Doc - Rust snippets moved in tests
